refactor(cli): split plugin command module [OPE-389]
